+++ test/test_deepen_activation.py
@@ -19,6 +19,22 @@ class DeepenActivationReluTest(unittest.TestCase):
 
         self.assertTrue(np.array_equal(cache, self.Z))
 
+class DeepenActivationReluBackwardTest(unittest.TestCase):
+    def setUp(self):
+        self.Z = np.array([[-0.8], [0], [0.9]])
+        self.dA = np.ones((3, 1))
+        self.dZ_expected = np.array([[0], [0], [1]])
+
+    def test_dZ_has_the_correct_shape(self):
+        dZ = activation.relu_backward(self.dA, self.Z)
+
+        self.assertTrue(dZ.shape == self.dZ_expected.shape)
+
+    def test_returns_the_gradient_of_the_input(self):
+        dZ = activation.relu_backward(self.dA, self.Z)
+
+        self.assertTrue(np.array_equal(dZ, self.dZ_expected))
+
 class DeepenActivationSigmoidTest(unittest.TestCase):
     def setUp(self):
         self.Z = np.array([[1], [2], [3]])